Implementação no de força resultante em módulo funcionando
