+++ lista.h
@@ -28,6 +28,9 @@ Lista *InsereForcaLista(Lista *l, float valor);
 void DesalocarListaForca(Lista *l);
 float SomaListaForca(Forcas *vet);
 void ImprimirListaForca(Forcas *list);
+int BuscarForca(Forcas *f, int posicao);
+void RemoverForca(Lista *l, int posicao);
+// float SubtrairLista(Lista *adc, Lista *sub);
 
 // IMPLEMENTAÇÕES
 // Inializa a lista como nula
@@ -37,7 +40,7 @@ Lista *InicializaListaForca(){
     return l;
 }
 
-// insere as forças na lista de forças
+// Insere as forças na lista de forças
 Lista *InsereForcaLista(Lista *l, float valor){
     Forcas *novo = (Forcas *)malloc(sizeof(Forcas));
     novo->forca = valor;
@@ -45,7 +48,8 @@ Lista *InsereForcaLista(Lista *l, float valor){
     l->inicio = novo;
     return l;
 }
-// Desaloca lista de forças
+
+// Desaloca lista de forças na memória
 void DesalocarListaForca(Lista *l){
     Forcas *aux = l->inicio; 
     while (aux != NULL){
@@ -56,7 +60,7 @@ void DesalocarListaForca(Lista *l){
     free(l);
 }
 
-// soma o modulo das forças e retorna o valor da força resultante
+// Soma os valores das forças em módulo e retorna o valor da força resultante
 float SomaListaForca(Forcas *vet){
     float soma = 0;
     Forcas *aux = vet;
@@ -67,12 +71,85 @@ float SomaListaForca(Forcas *vet){
     return soma;
 }
 
+// Imprime a lista de forças
 void ImprimirListaForca(Forcas *list){
     Forcas *aux = list;
+    int indice = 1;
     while(aux != NULL){
-        printf("Forca: %.10f N\n", aux->forca);
+        printf("Forca %d: %.10f N\n", indice, aux->forca);
         aux = aux->prox;
+        indice++;
     }
 }
+
+// Função que verifica se existe o valor da força na lista através da posição
+int BuscarForca(Forcas *f, int posicao){
+    Forcas *aux = f;
+    float valor = 0;
+    int cont = 1; // Posicão incial da lista
+    while (aux != NULL){
+        if (cont == posicao){
+            valor = aux->forca;
+            return valor; // Retorna o valor da força se encontrar na lista
+        }else{
+            aux = aux->prox;
+            cont++;
+        }
+    }
+    return 1; // Retorna 1 se não encontrar o valor da força na lista
+}
+
+// Função que remove o valor da força da lista
+void RemoverForca(Lista* l, int posicao) {
+    Forcas* atual = l->inicio;
+    Forcas* anterior = NULL;
+    float valor = BuscarForca(l->inicio, posicao);
+    int cont = 1; // Posicão incial da lista
+    
+    //Se o valor não estiver na lista de forças
+    if(valor == 1){
+        printf("Forca nao encontrada, tente novamente\n");
+        return ;
+    //Se o valor da força estiver na lista de forças
+    }else{ 
+        //se o primeiro nó for o nó a ser removido
+        while(anterior == NULL && cont == posicao){
+            l->inicio = atual->prox;
+            cont++;
+            free(atual);
+            return;
+        }
+        //se o nó a ser removido não for o primeiro
+        while(atual != NULL && cont != posicao){
+            anterior = atual;
+            atual = atual->prox;
+            cont++;
+        }
+        //se o nó a ser removido for encontrado
+        while(atual != NULL && cont == posicao){
+            anterior->prox = atual->prox;
+            free(atual);
+            return;
+        }
+    }
+}
+
+// // Função que subtrai os valores das forças de duas listas
+// float SubtrairLista(Lista *adc, Lista *sub){
+//     float subtracao = 0;
+//     Forcas *aux_adc = adc;
+//     Forcas *aux_sub = sub;
+//     while(aux_adc != NULL && aux_sub != NULL){
+//         if(aux_adc->forca < aux_sub->forca){
+//             subtracao = aux_sub->forca - aux_adc->forca;
+//         }else{
+//             subtracao = aux_adc->forca - aux_sub->forca;
+//         }
+        
+//         aux_adc = aux_adc->prox;
+//         aux_adc = aux_adc->prox;
+//     }
+//     return subtracao;
+// }
 
 #endif  // INCLUDE_LISTA_H_
