+++ tempCodeRunnerFile.c
@@ -1,0 +1,11 @@
+
+        MenuFisica();
+        scanf("%d", &r);
+        switch (r){
+        case 1:
+            ModForcaEletrostatica();
+            break;
+        case 2:
+            printf("Ainda nao desenvolvido\n");
+            break;
+        case 3:
